Refactor: Move x86 JIT support functions into C file
