+++ vm/CMakeLists.txt
@@ -44,7 +44,6 @@ add_library("ubpf"
   ubpf_jit_support.c
   ubpf_jit_support.h
   ubpf_jit_x86_64.c
-  ubpf_jit_x86_64.h
   ubpf_loader.c
   ubpf_vm.c
 )
